requirements.txt
pydantic==2.9.2
streamlit==1.39.0
requests==2.32.3

# Exporter optional dependencies (pinned to match backend)
python-docx==1.1.0
matplotlib==3.8.0
fpdf2==2.7.9
Pillow==10.1.0
